fix: move the SuiteRoutereHistory defitions from the redux-utils to the destkop package
